Fixed GO Term analysis, in that the background set is not ~700 proteins from the expanded set for the full yeast quantitative dataset. Reorganized some of the classes. More testing for GO Terms.
