+++ Test/TestTheoreticalCreation.cs
@@ -1,0 +1,34 @@
+﻿using NUnit.Framework;
+using ProteoformSuiteInternal;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using Proteomics;
+using System.Linq;
+
+namespace Test
+{
+    [TestFixture]
+    public class TestTheoreticalCreation
+    {
+        [Test]
+        public void test_get_modification_dictionary()
+        {
+
+        }
+
+        [Test]
+        public void test_protein_grouping_by_sequence()
+        {
+            DatabaseReference d1 = new DatabaseReference("GO", ":", new List<Tuple<string, string>> { new Tuple<string, string>("term", "P:") });
+            DatabaseReference d2 = new DatabaseReference("GO", ":", new List<Tuple<string, string>> { new Tuple<string, string>("term", "P:") });
+            DatabaseReference d3 = new DatabaseReference("GO", ":", new List<Tuple<string, string>> { new Tuple<string, string>("term", "P:") });
+            GoTerm g1 = new GoTerm(d1);
+            GoTerm g2 = new GoTerm(d1);
+            GoTerm g3 = new GoTerm(d1);
+            ProteinWithGoTerms p1 = new ProteinWithGoTerms("MSSSSSSSSSSS", "T1", new Dictionary<int, List<Modification>>(), new int?[] { 0 }, new int?[] { 0 }, new string[] { "" }, "T2", "T3", true, false, new List<DatabaseReference> { d1 }, new List<GoTerm> { g1 });
+            ProteinWithGoTerms p2 = new ProteinWithGoTerms("MSSSSSSSSSSS", "T2", new Dictionary<int, List<Modification>>(), new int?[] { 0 }, new int?[] { 0 }, new string[] { "" }, "T2", "T3", true, false, new List<DatabaseReference> { d2 }, new List<GoTerm> { g2 });
+            ProteinWithGoTerms p3 = new ProteinWithGoTerms("MCSSSSSSSSSS", "T3", new Dictionary<int, List<Modification>>(), new int?[] { 0 }, new int?[] { 0 }, new string[] { "" }, "T2", "T3", true, false, new List<DatabaseReference> { d3 }, new List<GoTerm> { g3 });
+        }
+    }
+}
